+++ module-services/service-appmgr/messages/CloseConfirmation.cpp
@@ -1,0 +1,11 @@
+// Copyright (c) 2017-2020, Mudita Sp. z.o.o. All rights reserved.
+// For licensing, see https://github.com/mudita/MuditaOS/LICENSE.md
+
+#include <service-appmgr/messages/CloseConfirmation.hpp>
+
+namespace app::manager
+{
+    CloseConfirmation::CloseConfirmation(const ApplicationName &sender)
+        : BaseMessage(MessageType::APMConfirmClose, sender)
+    {}
+} // namespace app::manager